backend/requirements.txt
fastapi==0.95.1
uvicorn==0.23.0
web3==6.0.1
requests==2.28.1
python-dotenv==0.21.0
